FileListing class to handle single directory listings for 1.x branch (#910)
* Read 'ls -la /directory' single directory output

Jakub Svoboda identified in issue #827 that the FileListing parser wasn't
able to read single directory listings, because a single directory listing
doesn't include the directory name in the output and only if the directory
name is read will anything else be captured.  This patch fixes that by
checking for a path on the command line or in the file path in sosreport
format - e.g. 'ls_-la_.etc.pki.tls', with '.' substituted for '/'.  That
dotted suffix, converted back to path format, is used as a 'first guess' of
the directory.  If a directory heading is found in the output that is used
in preference to the first guess.

* Make sure first path guess is ignored once a path is read

* Change '_' to ' ' in directory paths, and add caveat on single dir usage