Event stream handling
Includes a new `stream` component type and `Stream` base class.

A `Stream` is a `threading.Thread` with an input queue. Standard dependency
injection is used to attach component input queues to their dependencies. If a
dependency isn't a `Stream` object, we don't attempt to wire it up.

`Stream` subclasses may implement __init__ to inject non `Stream` dependencies
but should pass all `Stream` dependencies on to the superclass __init__.

`Stream` subclasses should implement `update(self, source, event)` as this
function gets called whenever an upstream dependency emits an event. `source`
is the class of the dependency, and `event` is the object it emitted. `event`
will never been the same class as source. Instead, components should emit
objects relevant to what they represent.

`run_streams` is the entry point and has the same signature as `dr.run`. It
accepts a graph of components that defaults to all loaded `stream`s. It also
accepts a broker that you can seed with initial objects. A blank broker is
created if one isn't provided.