Updates to cat and spec_factory for multioutput datasources.
Give insights/tools/cat.py the ability to output multiple datasources
in a way that makes them easy to break back apart. Also, create a new
datasource helper function in spec_factory that knows how to read files
created in that way with cat. Create a new CatDataProvider class to surface
the content to parsers.